# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-tenant video platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /permission-groups:
    get:
      operationId: listPermissionGroups
      tags: [members]
      responses:
        "200":
          description: Permission groups available in the current workspace
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionGroup"
    post:
      operationId: createPermissionGroup
      tags: [members]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionGroupInput"
      responses:
        "201":
          description: Permission group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionGroup"
  /permission-groups/{groupId}:
    patch:
      operationId: updatePermissionGroup
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionGroupInput"
      responses:
        "200":
          description: Permission group updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionGroup"
    delete:
      operationId: deletePermissionGroup
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "204":
          description: Permission group deleted
...
  /members/{membershipId}:
    patch:
      operationId: updateMember
      tags: [members]
      parameters:
        - $ref: "#/components/parameters/MembershipId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /invitations:
    post:
      operationId: createInvitation
      tags: [members]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationInput"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
...
components:
  parameters:
...
    GroupId:
      in: path
      name: groupId
      required: true
      schema:
        type: string
    MembershipId:
      in: path
      name: membershipId
      required: true
      schema:
        type: string
  schemas:
...
    PermissionGroup:
      type: object
      required: [id, name, description, permissions]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        permissions:
          type: array
          items: { type: string }
    PermissionGroupInput:
      type: object
      required: [name, description, permissions]
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        description: { type: string, maxLength: 240 }
        permissions:
          type: array
          uniqueItems: true
          items: { type: string }
...
    Member:
      type: object
      required: [id, name, email, groupId, role, status]
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, pattern: "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$" }
        groupId: { type: string }
        role: { type: string }
        status:
          type: string
          enum: [invited, active, suspended]
    MemberUpdate:
      type: object
      properties:
        groupId: { type: string }
        status:
          type: string
          enum: [active, suspended]
    InvitationInput:
      type: object
      required: [email, groupId]
      properties:
        email: { type: string, pattern: "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$" }
        groupId: { type: string }
    Invitation:
      type: object
      required: [id, email, groupId, role, status, expiresAt]
      properties:
        id: { type: string }
        email: { type: string, pattern: "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$" }
        groupId: { type: string }
        role: { type: string }
        status:
          type: string
          enum: [invited]
        expiresAt: { type: string, format: date-time }
...